backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
pillow==11.3.0
pydantic==2.11.10
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
httpx
python-dotenv==1.0.1
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
